requirements.txt
# LangChain core packages
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.2.0

# For compatibility with classic LangChain chains
langchain>=0.2.0

# Vector database
chromadb>=0.4.0

# Audio transcription
openai-whisper>=20230918

# PDF processing
pypdf>=3.0.0
